+++ data/src/main/java/ru/radiationx/data/entity/response/release/ExternalEpisodeResponse.kt
@@ -5,7 +5,7 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class ExternalEpisodeResponse(
-    @Json(name = "id") val id: Int,
+    @Json(name = "id") val id: Float,
     @Json(name = "title") val title: String?,
     @Json(name = "url") val url: String?
 )
